[refactor] move HUD class, HTML and functionality out to separate file
